Adding test websocket for receiving and sending messages v4
