--- src/main/java/services/UserFounderService.java
@@ -1,7 +1,0 @@
-package services;
-
-import models.Festival;
-
-public interface UserFounderService {
-    void isUserInFestival(String point, Festival festival);
-}
