+++ src/main/java/servlets/LocationWebSocketServlet.java
@@ -2,36 +2,40 @@ package servlets;
 
 
 
+import models.Festival;
 import services.LocationWebSocketConfigurator;
+
 
 import javax.websocket.*;
 import javax.websocket.server.ServerEndpoint;
-import java.util.Collections;
-import java.util.HashSet;
-import java.util.Set;
+import java.util.*;
 
 @ServerEndpoint(value = "/compareLocations", configurator= LocationWebSocketConfigurator.class)
 public class LocationWebSocketServlet {
+        String point;
+        Festival festival = new Festival("Test", "Testing", "Red" ,
+            "Some", "60.11173060613703 30.267900556923905", 75);
+
 
         private Set<Session> userSessions = Collections.synchronizedSet(new HashSet<Session>());
 
         @OnOpen
         public void start(Session userSession) {
+            System.out.println("Connected user:" + userSession.getId());
             userSessions.add(userSession);
         }
 
         @OnClose
         public void onClose(Session userSession) {
             userSessions.remove(userSession);
+            System.out.println("Connected user:" + userSession.getId());
         }
 
         @OnMessage
-        public void onMessage(String message, Session userSession) {
+        public void onMessage(String message, Session userSession) throws Throwable {
             System.out.println("Message Received: " + message);
-            for (Session session : userSessions) {
-                System.out.println("Sending to " + session.getId());
-                session.getAsyncRemote().sendText(message);
-            }
+            point = message;
+            sendRequestToUpdate(userSession);
         }
 
         @OnError
@@ -39,9 +43,38 @@ public class LocationWebSocketServlet {
             //TODO
         }
 
+        public void sendRequestToUpdate(Session session) throws Throwable {
+            Timer timer = new Timer();
+            timer.schedule(new TimerTask() {
+                @Override
+                public void run() { session.getAsyncRemote().sendText("getCoordinates from"+session.getId());
+                }
+            },10000);
+            boolean b = isInUnit(point,festival);
+            System.out.println(b + " "+ session.getId());
+        }
 
-        public void sendRequestToUpdate(Throwable t) throws Throwable {
 
+        private boolean isInUnit(String point, Festival festival) {
+            double[] pointCoordinates = getCoordinates(point);
+            double[] festivalCoordinates = getCoordinates(festival.getCenter());
+            double dx = pointCoordinates[0] - festivalCoordinates[0];
+            double dy = pointCoordinates[1] - festivalCoordinates[1];
+            double dx2 = Math.pow(dx, 2);
+            double dy2 = Math.pow(dy, 2);
+            double sum = dx2 + dy2;
+            double sqrt = Math.sqrt(sum);
+            double meters = sqrt*100000;
+            return meters - festival.getRadius()<=0;
+        }
+
+        private double[] getCoordinates(String center) {
+            double[] coordinates = new double[2];
+            String[] XY = center.split(" ");
+
+            coordinates[0] = Double.parseDouble(XY[0]);
+            coordinates[1] = Double.parseDouble(XY[1]);
+            return coordinates;
         }
