+++ src/main/webapp/userNotification.js
@@ -105,9 +105,10 @@ function sendMessage(webSocketClient) {
         navigator.geolocation.getCurrentPosition(function (position) {
             userX = position.coords.latitude;
             userY = position.coords.longitude;
-            var message = userX + ";" + userY;
-            var jsonObj = {"x" : userX, "y" : userY};
-            webSocketClient.send(JSON.stringify(jsonObj));
+            var message = userX + " " + userY;
+            //var jsonObj = {"x" : userX, "y" : userY};
+            //webSocketClient.send(JSON.stringify(jsonObj));
+            webSocketClient.send(message);
         });
 }
 
@@ -117,11 +118,11 @@ function connect() {
     webSocketClient.onopen = function (event) {
         console.log("onopen " );
         sendMessage(webSocketClient);
-        }
+        };
     webSocketClient.onmessage = function (event) {
             console.log("event.data " + event.data);
-
-        }
+            sendMessage(webSocketClient);
+        };
     webSocketClient.onclose = function (event) {
             console.log("close");
